refactor(memex): enforce verified outcome module bounds
Worker-Lane: memex-runtime-authority

Slice: FOUNDUP_MEMEX_VERIFIED_OUTCOME_RUNTIME_AUTHORITY_BINDING_PHASE1
