+++ extensions/reddog/backend_compatibility_constants.js
@@ -4,7 +4,7 @@ const BACKEND_MANIFEST_SCHEMA = 'reddog_backend_manifest.v3';
 const BACKEND_PRODUCT = 'foundups-agent-reddog-backend';
 const BACKEND_API_VERSION = 2;
 const BACKEND_MANIFEST_PATH = 'scripts/reddog_backend_manifest.json';
-const EXPECTED_MANIFEST_SHA256 = '9b0cc050089af50c31ecb42a3d10f5e85dccd3339d8543ee34c86eaef6f75a6a';
+const EXPECTED_MANIFEST_SHA256 = 'dff7b6bcf45d9bebefe78fe4ad23dd73ccd2b0902e1284753c61a36c612bb38d';
 const RUNTIME_DEPENDENCY_GRAPH_VERSION = 2;
 const MAX_MANIFEST_BYTES = 256 * 1024;
 const MAX_RUNTIME_FILE_BYTES = 2 * 1024 * 1024;
