+++ modules/communication/moltbot_bridge/tests/test_reddog_main_openclaw_signed_worker_claim_loop_preflight.py
@@ -884,7 +884,7 @@ def test_main_openclaw_signed_0102_bounded_code_uses_fusion_artifact_generation(
     assert "REDDOG_ARTIFACT_CONTENTS_PATH" not in os.environ
 
 
-def test_main_openclaw_signed_worker_claim_loop_completes_env_bound_chain(
+def test_main_openclaw_signed_worker_claim_loop_stops_before_unauthorized_pattern_memory(
     tmp_path: Path,
     monkeypatch,
     capsys,
@@ -1074,13 +1074,14 @@ def test_main_openclaw_signed_worker_claim_loop_completes_env_bound_chain(
         )
 
     captured = capsys.readouterr().out
-    assert "[REDDOG-OPENCLAW-CLAIM-LOOP] preflight=PASS" in captured
-    assert "status=SIGNED_WORKER_OPENCLAW_CLAIM_LOOP_ACCEPT" in captured
+    assert "[REDDOG-OPENCLAW-CLAIM-LOOP] preflight=WARN" in captured
+    assert "status=SIGNED_WORKER_OPENCLAW_CLAIM_LOOP_REJECT" in captured
     assert "claimed_count=7" in captured
+    assert "FAIL_HANDLER_MISSING,stage:pattern_memory_admission" in captured
     assert coding_task_id in captured
     assert queue_stage_task_id in captured
     assert AgentDB().get_autonomous_task_by_id(coding_task_id)["status"] == "completed"
-    assert AgentDB().get_autonomous_task_by_id(queue_stage_task_id)["status"] == "completed"
+    assert AgentDB().get_autonomous_task_by_id(queue_stage_task_id)["status"] == "failed"
     remaining_signed = [
         task
         for task in AgentDB().get_autonomous_tasks(status="pending", limit=10)
@@ -1098,13 +1099,12 @@ def test_main_openclaw_signed_worker_claim_loop_completes_env_bound_chain(
         "verified_outcome_ratchet",
         "model_feedback_admission",
         "held_out_regression_gate",
-        "pattern_memory_admission",
     ):
         assert stage_name in stored["stage_results"]
+    assert "pattern_memory_admission" not in stored["stage_results"]
     generation = stored["stage_results"]["bounded_worker_pilot"]["artifact_generation_result"]
     assert generation["accepted"] is True
     assert generation["receipt"]["model_receipt_id"] == "fusion-artifact-receipt-1"
-    assert stored["receipts"][-1]["next_action"] == "STOP_QUEUE_CHAIN_COMPLETE"
     assert (worktree / PILOT_ARTIFACT).read_text(encoding="utf-8").startswith(
         "# Generated By Fusion"
     )
@@ -1117,9 +1117,7 @@ def test_main_openclaw_signed_worker_claim_loop_completes_env_bound_chain(
     ]
     assert draft_pr_calls == ["commit_all", "push_branch", "create_draft_pr"]
     assert outcome_store.exists()
-    with sqlite3.connect(pattern_memory_db) as conn:
-        count = conn.execute("SELECT COUNT(*) FROM skill_outcomes").fetchone()[0]
-    assert count == 1
+    assert not pattern_memory_db.exists()
     assert not (repo / "runtime" / "pattern_memory.db").exists()
