+++ scripts/reddog_backend_manifest.json
@@ -1526,7 +1526,7 @@
     "modules/communication/moltbot_bridge/src/foundup_brain_current_state.py": "82f6453086a8bff9238795bc5580d427f43753b40e508081cfd405ba379f79cd",
     "modules/communication/moltbot_bridge/src/foundup_job_contract.py": "fc3fead2e3cb381b9b848b642d8ececfa73027d66dd8ca6a31a4819e26952693",
     "modules/communication/moltbot_bridge/src/foundup_memex_current_state.py": "b7a03af9fa2d1baa1b3edb3cb17c62759a3696d5c5f3afa853b66338501502fe",
-    "modules/communication/moltbot_bridge/src/foundup_memex_verified_outcome_authenticity.py": "1d53f3e0f6a124d90019fd3affebb5d4a7527547615a59a0faa41a50b0aa4195",
+    "modules/communication/moltbot_bridge/src/foundup_memex_verified_outcome_authenticity.py": "4c40dac1b6f4714ba6222d401071fcb41bbb93e148a785c2c0e39003ceadd0c4",
     "modules/communication/moltbot_bridge/src/foundup_memex_verified_outcome_publisher.py": "09c8d15e03cb9136e3f86af638ebe6c0de80f90358deada9e1107e15aeb8a0fc",
     "modules/communication/moltbot_bridge/src/foundup_memex_verified_outcome_queue_binding.py": "a22f95a22da7d3939929fcf56561d404d985646ce3a0ecb08b68863db48b060d",
     "modules/communication/moltbot_bridge/src/foundup_memex_verified_outcome_receipt_rehydration.py": "885257c146e1066911d800406b56836e96e5f26c2f5aa1c611f12de1c6bdd5ba",
